+++ config/settings/base.py
@@ -123,8 +123,8 @@ INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
 # ------------------------------------------------------------------------------
 # https://docs.djangoproject.com/en/dev/ref/settings/#middleware
 MIDDLEWARE = [
-    "safe_transaction_service.middlewares.LoggingMiddleware",
-    "safe_transaction_service.middlewares.ProxyPrefixMiddleware",
+    "safe_transaction_service.middlewares.logging_middleware.LoggingMiddleware",
+    "safe_transaction_service.middlewares.proxy_prefix_middleware.ProxyPrefixMiddleware",
     "corsheaders.middleware.CorsMiddleware",
     "django.middleware.security.SecurityMiddleware",
     "django.contrib.sessions.middleware.SessionMiddleware",
